Resolve memory.py dual allegiance: rename to context_assembler.py, leave shim
memory.py was doing prompt/context assembly, not memory governance.
Renamed to context_assembler.py to tell the truth about its role.
Class renamed MemoryAssembler -> ContextAssembler with backward compat alias.
memory.py becomes a thin shim that re-exports the public symbols.
README updated: context_assembler.py moves to substrate, memory.py demoted
to transitional, *_types.py explicitly declared as codebook schema stratum.
